GoSet1.go
Here is the Set 1 that I have answered utilizing the Go Language

Sadly, the online checker does not work for me because I approached the code using multiple functions and calling it with one another. I believe that this has caused the checker to not being able to read the file properly that led to issues on verifying if the code is working. 

As an alternative, ive ran multiple tests on my own and I believe that these functions are working